feat(telephony): implement SIM subscription information gathering
Introduce a wrapper around the SubscriptionManager to grab and
reorganize information of all available SIM subscriptions found on the
device.

The usage is simple and boils down to injecting the Subscriptions class
wherever it's needed. For now, there are two implementations of the
Subscriptions class - one for devices (except eUICC ones) supporting the
newer Radio Interface Layer, the other for all other devices only
supporting the legacy Radio Interface Layer. The implementation to be
used will be determined at runtime during injection, depending on the
supported RIL by the device.

All in all, this is a "MUST-HAVE" for our business model (coming next),
as it will allow to simplify and reduce boilerplate code instead of
directly using the SubscriptionManager's APIs in the application's
codebase, also facilitate testing and understand better what's going on
since we're using system APIs that sometimes lack documentation.

Signed-off-by: iusmac <iusico.maxim@libero.it>