build(Gradle): bump `androidx.core` to `1.9.0-alpha05`
Build fails because Android T (API 33) sources still ship the library
AndroidX Core v1.9.0-alpha03, which is a preview release labeled with
'Tiramisu' as its codename.

To fix it, we're going to upgrade the library to v1.9.0-alpha05, which,
as per CL in [1], only aligns the compile version with SDK 33 without
any other code change.

Ref:
  [1] https://developer.android.com/jetpack/androidx/releases/core#1.9.0-alpha05

----

FAILURE: Build failed with an exception.
* What went wrong:
    Execution failed for task ':app:checkDebugAarMetadata'.
    > An issue was found when checking AAR metadata:
        1.  Dependency 'androidx.core:core:1.9.0-alpha03' requires libraries and applications that
            depend on it to compile against codename "Tiramisu" of the
            Android APIs.
:app is currently compiled against android-33.
Recommended action: Use a different version of dependency 'androidx.core:core:1.9.0-alpha03', or set compileSdkPreview to "Tiramisu" in your build.gradle file if you intend to experiment with that preview SDK.

Signed-off-by: iusmac <iusico.maxim@libero.it>