feat(telephony): implement pSIM toggle support
Prepare a wrapper around the TelephonyManager and SubscriptionManager to
to make it easier to manage the enabled state of available SIM cards and
their subscriptions found on the device.

The usage is simple and boils down to injecting the TelephonyController
or SubscriptionController classes whenever they're needed.

The TelephonyController implementation is targeting devices that only
support the legacy Radio Interface Layer (RIL). However, it can be used
by devices that also support the newer RIL as the default method to
control the enabled state of SIM cards.

The SubscriptionController implementation is targeting only devices
(except eUICC ones) that support the newer Radio Interface Layer.

Signed-off-by: iusmac <iusico.maxim@libero.it>