feat(ui/sim): add activity to list all available SIM cards on the device
List all available SIMs on the device as preferences.

Each SIM preference has an icon that is tinted with the color selected
in the system built-in Setting app.

Each SIM provides a switch allowing the user to toggle its enabled
state.

The SIM list is live; it is automatically updated on SIM insertion /
removal, or when any of the SIM attributes (for instance name or tint
color) have changed.