feat(AndroidManifest): require this app for all users when creating a restricted profile
Signal to built-in system Settings app that this app is required for all
users when a restricted profile is created.

See this plausible example: User 1 powers down a SIM card. Subsequently,
a switch to the User 2 occurred. Since SIM cards are not tied to a
specific user, but rather to the device, the User 2 will end up with the
physical SIM card in the SIM card tray slot, but completely missing in
the system and unable to turn it on because Android doesn't provide such
a feature out-of-the-box.
Therefore, the application should always be available.

Signed-off-by: iusmac <iusico.maxim@libero.it>