build(Gradle): bump `androidx.appcompat` to `1.6.0-alpha05`
Build fails because the Android T (API 33) sources provide the preview
release of appcompat v1.6.0-alpha02 library, which doesn't exist in
public repositories.

To fix it, we're going to upgrade the library to v1.6.0-alpha05, which,
as per CL in [1], only aligns the compile version with SDK 33 without
any code additions/removals.

Ref:
    [1] https://developer.android.com/jetpack/androidx/releases/appcompat#1.6.0-alpha05

---

* What went wrong:
Execution failed for task ':app:checkDebugDuplicateClasses'.
    > Could not resolve all files for configuration ':app:debugRuntimeClasspath'.
    > Could not find androidx.appcompat:appcompat:1.6.0-alpha02.
Searched in the following locations:
    - https://dl.google.com/dl/android/maven2/androidx/appcompat/appcompat/1.6.0-alpha02/appcompat-1.6.0-alpha02.pom
    - https://repo.maven.apache.org/maven2/androidx/appcompat/appcompat/1.6.0-alpha02/appcompat-1.6.0-alpha02.pom
Required by:
    project :app > project :SettingsLib:CollapsingToolbarBaseActivity

Signed-off-by: iusmac <iusico.maxim@libero.it>