perf(ui/sim): add debouncing for the SIM list updates
Under the hood, we're listening for the SubscriptionManager, which fires
whenever a subscription field changes. This causes the UI thread to
block, resulting in noticeable UI stuttering.

A proper solution to this would be accumulating multiple changes to
affected subscriptions in a batch and fire just once. Interestingly, the
SIM detail activity in the system built-in Settings app is also affected.

As a workaround, we can try to "calm down" the SubscriptionManager on
our side by delaying and discarding all redundant updates that happen
within 300ms. This will simulate a bulk update and make the life easier
for the UI thread.