chore: introduce `ApplicationInfo`
Since this is a FOSS application that is supposed to be integrated with
any ROM, it will be signed using public AOSP platform signature by
default. Nevertheless, the maintainer can also prefer to sign with its
own private keys (aka dev-keys signature). Therefore, we need to prepare
a convenient API to verify whether the application has been signed with
the AOSP platform signature, in order to protect users from taking
actions that might lead to a broken application due to signature
mismatch.

Note that, from a performance perspective, we determine the value on app
startup and cache the result within the Application class, which is a
singleton that exists for the lifetime of the app process.

Signed-off-by: iusmac <iusico.maxim@libero.it>