feat(preferences): add launcher icon visibility setting
Feature stolen from GBoard keyboard but improved:

- the application won't close after hiding its icon (fixed by using task
affinity and a proxy launcher activity).

- the preference remains visible but disabled in the preference list
instead of removing it, ensuring users are aware of its existence even
if the application isn't installed in the device's system partition
(as of Android Q only system apps are allowed to hide launcher icon).

- the user's preference won't be reset if the launcher icon was hidden
and the application was converted from a system to a user app (icon
visibility must be restored as of Android Q), then if converted back
to system app, the user's preference will be kept and launcher icon
will be hidden again on boot completed. This also ensures proper app
backup data restore.

Signed-off-by: iusmac <iusico.maxim@libero.it>