tweak(launcher): improve app icon hiding approach
Previously, to hide the launcher icon in the launcher's app list, we
involved an invisibile activity that launched the main activity and
self-closed afterwards. This was necessary to split the launcher
activity from the main activity stack, and allow disabling it without
forcing the system to close the app. The downside of the this approach
is the app startup slowdown due the extra activity being launched before
the UI was displayed.

This commit aims to address app startup time regression introduced by
the old approach, which boils down to postponing launcher icon hiding
until the app is about to reach the end of its lifecycle, eliminating
the need for an extra activity.

Signed-off-by: iusmac <iusico.maxim@libero.it>