feat(ui): implement Toolbar decorator
With the toolbar decorator we can now apply the marquee effect for the
framework Toolbar's title/subtitle, or, if there's support for
collapsing toolbar, decorate it with a collapsing subtitle, which isn't
supported out of the box.

By default, the new version of the CollapsingToolbarBaseActivity will:
- apply the marquee effect on the title for non-collapsing Toolbar
  (reason: for better UX)
- override the default AOSP's value of 50ms for scrim's in/out animation
  (reason: too short and makes the scrim flicker on 60Hz displays)

While we're at it, move all activities to the new version of the
CollapsingToolbarBaseActivity.

Signed-off-by: iusmac <iusico.maxim@libero.it>