chore(deps): add prebuilt `android-33.jar` as compile-only dependency
Since this app leverages Android hidden APIs, let's ease the life
for whoever wants to build it just by invoking 'gradle build'.

We may also necessitate for it in automated CI builds.

NOTE: although the build passes, for the development it's still required
to manually copy the library to the Android SDK directory in order to
avoid the "Cannot resolve symbol" error. It looks like the framework
library appears at the end of the classpath list and, only the first
occurrence (i.e the stock 'android.jar' file) is used to resolve
symbols. In the javac 1.8 and older, this process was automated by
overriding the default bootstrap classpath. Currently, there's doesn't
seem to be any other way to workaround it...

Signed-off-by: iusmac <iusico.maxim@libero.it>