ci(release-please): auto-generate `Contributors` for the latest release
This commit introduces a fully-automated generation of the contributors
list that will be appended to the end of the release summary when
creating the latest GitHub release using Release Please action.

All work is done by the 'add-release-contributors.sh' script that relies
on the GitHub CLI tools to make relative API calls. It can also be used
outside of CI context to add Contributors block to any release (except
the very first one) in any repository, until the 'gh' command has the
PAT (Private Access Token).

Signed-off-by: iusmac <iusico.maxim@libero.it>