feat(telephony/legacy): implement persistent SIM card disabled state
As per TelephonyController specs, turning off a SIM card on a device
using legacy RIL won't persist across boots. On reboot, SIM will turn on
normally, but the user may want it to be turned off until turned on again
manually or through a weekly repeat schedule for that SIM card, if any.

Signed-off-by: iusmac <iusico.maxim@libero.it>