build(Gradle): bring back old `gitDescribeTag()` but improve constraints
The 33b6122d32295c0165ff8f8af467dac29b72d75a actually introduces a
regression. If using this function in a project that doesn't have tags
at all, the 'git describe' command will return just "commit-ish", but we
expect at least 3 pieces, so this case must be handled as it was
intended before.

This commit brings back the old implementation, but we keep the --long
argument and rewrite the constraints to avoid 'index out of range'
errors.

Signed-off-by: iusmac <iusico.maxim@libero.it>