test(UserAuthenticationObserverServiceTest): attempt to fix test flakiness
When running this particular test inside a GitHub workflow, for some
reason, sending the interrupt signal can take some time to be received
by the Worker thread, so the LinkedBlockingQueue poll is not stopped and
new tasks can still be "infiltrated", or, maybe the latter itself takes
time to acquire a new ReentrantLock and enter the waiting state.
Luckily, we have the mReleased flag, which we can use to further prevent
new tasks from being added to the queue when Worker has been explicitly
stopped.

java.lang.AssertionError:
Expected: is <3>
     but: was <4>
	at org.hamcrest.MatcherAssert.assertThat(MatcherAssert.java:20)
	at org.hamcrest.MatcherAssert.assertThat(MatcherAssert.java:6)
	at com.github.iusmac.sevensim.UserAuthenticationObserverServiceTest$TaskRunningLifecycle.test_onStartCommand_NoNewTasksShouldBeAcceptedAfterDestroyed(UserAuthenticationObserverServiceTest.java:365)

Signed-off-by: iusmac <iusico.maxim@libero.it>