test(ForegroundServiceTest): attempt to address timeout in a "white box" test
This is an integration test that involves multi-threading, so we wait
for it to complete within 2 seconds (normally, it supposed to complete
under 1s). But when running the same test on GitHub-hosted runners, it
seems that there's CPU throttling or different threads run on the same
physical CPU core, which makes the newly spawned threads start late and
take longer to finish (again due to throttling and tweaked scheduling).

As a workaround, we'll just wait another second for the task to
complete.

---

org.awaitility.core.ConditionTimeoutException: Condition with Lambda expression in com.github.iusmac.sevensim.ForegroundServiceTest was not fulfilled within 2 seconds.
	at org.awaitility.core.ConditionAwaiter.await(ConditionAwaiter.java:167)
	at org.awaitility.core.CallableCondition.await(CallableCondition.java:78)
	at org.awaitility.core.CallableCondition.await(CallableCondition.java:26)
	at org.awaitility.core.ConditionFactory.until(ConditionFactory.java:1160)
	at org.awaitility.core.ConditionFactory.until(ConditionFactory.java:1129)
	at com.github.iusmac.sevensim.ForegroundServiceTest.awaitServiceStoppedBySelfWithinTime(ForegroundServiceTest.java:1099)
	at com.github.iusmac.sevensim.ForegroundServiceTest.awaitServiceStoppedBySelfWithinTime(ForegroundServiceTest.java:1105)
	at com.github.iusmac.sevensim.ForegroundServiceTest$Base.awaitStarted(ForegroundServiceTest.java:1066)
	at com.github.iusmac.sevensim.ForegroundServiceTest$UpdateNextWeeklyRepeatScheduleProcessingIterAction.test_WithDateTimeAndDecryptPinStorageFlagMatchingAllPinStorageEntities(ForegroundServiceTest.java:718)

Signed-off-by: iusmac <iusico.maxim@libero.it>