ToolbarDecorator: support auto-adjustable collapsing toolbar height when expanded
This removes the max 3 lines limit on the subtitle. It also improves the
way the toolbar title will position itself within the toolbar view, as
its gravity is set by the library to be at the bottom of the
CollapsingToolbarLayout, but we may have a subtitle at the very bottom
that also requires room, so they will overlap if the title is not pushed
up using setExpandedTitleMarginBottom() so that it's above the subtitle.

To work around the above, we continue to push the title to be above the
subtitle, but we also increase or shrink the overall height of the
CollapsingToolbarLayout, so that the title stays in the expanded area
and not behind the scrim.