fix(#82): ensure content is scrolled beneath toolbar (#83)
Fixes #82.

By default, the appbar starts as not lifted, then it's lifted on scroll.
That's why the scrollable content is not beneath the toolbar as seen in
#82. Our use case doesn't require appbar lifting, so we can safely
disable this feature, making the toolbar be permanently lifted above the
scrollable content.

This would likely be a bug in the material-components library, which
doesn't properly handle the edge-to-edge enforcement since Android 15.

NOTE: whereas disabling lifting feature fixes the view drawing order
(aka Z order), for a better UI, we also need to disable the scrim
background animation in Expressive Design (enabled by default since
Android 16), otherwise we'll see for a moment the scrollable content
through the AppBarLayout.

Signed-off-by: iusmac <iusico.maxim@libero.it>